pyproject.toml
[project]
name = "aimealplanner"
version = "0.1.0"
description = "Telegram meal planner bot bootstrap for the MVP"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.26.0",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "python-dotenv>=1.2.2",
    "redis>=7.3.0",
    "sqlalchemy>=2.0.48",
]

[project.scripts]
aimealplanner = "aimealplanner.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.6",
    "ty>=0.0.23",
]

[tool.hatch.build.targets.wheel]
packages = ["aimealplanner"]

[tool.ruff]
line-length = 100
target-version = "py313"
src = ["aimealplanner", "tests"]

[tool.ruff.lint]
select = ["ASYNC", "B", "E", "F", "I", "RUF", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
